+++ pext/qml/ModuleData.qml
@@ -192,7 +192,7 @@ Item {
                 visible: resultList.hasEntries && resultList.normalEntries == 0 && resultList.commandEntries == 0 && !searchInputFieldEmpty
 
                 text: "<h2>" + qsTr("No results") + "</h2>" +
-                      (resultList.unprocessedQueueCount > 0 ? ("<p>" + qsTr("Still processing %1 module request(s)…", "", resultList.unprocessedQueueCount) + "</p>") : "")
+                      (resultList.unprocessedQueueCount > 0 ? ("<p>" + qsTr("Still processing %1 module request(s)…".arg(resultList.unprocessedQueueCount), "", resultList.unprocessedQueueCount) + "</p>") : "")
 
                 anchors.centerIn: parent
                 color: palette.text